change movement to free movement
Up untill now, the movemnt was always restricted to a whole square. If you pressed a directional button, you moved a complete square.

This made animations feel sluggish, and gave the feeling that the movement did not really respond well to input.

For our use case, we don't actually need this restriction. Free movement works just as well.

Am added benefit is that it allows us to clean up the code and remove the coroutine.