add collision and global illumination
Configured the project to use Unity's Universal Render Pipelinie, along with 2D Lightning.

On top of that, enabled physics so that built in collisions work. This with a graviy of zero along both axes and frozen rotations: this prevents the sprite from "falling down" and from rotating around other physics objects.